merge updates from ms / master #4 (#4)
* Add WriteInPublic client

This is a very simple wrapper around the WriteInPublic API. Currently it
only implements the features that we need for integrating it into
Pombola.

I've added some tests for the basic functionality, but in the interests
of time I've skipped writing tests for the .people() and .answers()
methods, so will need to revisit that at some point in the future.

* Add WriteInPublic message creation wizard

This is a multi-step form wizard that walks a user through the steps of
creating a new message to a representative.

It uses django-formtools to roughly replicate the process of writing to
a representative as it currently exists on writeinpublic.com. Once the
user has written a message it uses the WriteInPublic API to send the
message and then takes the user to a confirmation page.

* Add view to display individual WriteInPublic messages

We want to be able to display messages from a WriteInPublic instance on
the site. In order to do this we need to retrieve messages from the
WriteInPublic API and then display them alongside the metadata for the
message.

* Add WriteInPublic view for displaying messages to a person

This uses the WriteInPublic API to pull back messages that have been
written to a specific person and displays them on a page.

This view isn't actually hooked into a URL yet, but the intention is to
AJAX it into the person profile page, I'll do the changes for that in a
separate commit/PR.

* [ZA] Styling for WriteInPublic integration

* [ZA] No need for @2x images any more

* Use chosen.js to enhance WriteInPublic recipients screen

This makes the dropdown the on the WriteInPublic recipients screen
support autocomplete, rather than making the user scroll through a long
list of names.

* [ZA] MVP ajax loading of message history into person page

* [ZA] Demote email address on profile page if WiP is an option

* [ZA] Activate WriteInPublic integration

* [ZA] Temporarily disable WriteInPublic on person profiles

The person profiles only show the links to write to that person if they
have an EP UUID, so temporarily change that method to always return none
while we get things setup.

* Revert "[ZA] Temporarily disable WriteInPublic on person profiles"

This reverts commit cbd77ff1d90f4b0928020bf142c3a431b458bed4.

* new homepage, basic mobile version and integration

* desktop layout for new homepage

* change navigation menu items

* homepage tweaks

- 'coming soon' button
- Mobile layout tweaks
- Buttons with icons

* Fix casing on "MPs"

* Add survey link

This is hard-coded for now, but we'll need to make this configurable in
the near future.

* [ZA] Fix featured MP on homepage

* [ZA] Update homepage links and Meet your MP section

* Change link on WriteInPublic message page

This now links back to the "Write to my MP" wizard, rather than being a
correct this page link.

This change was requested by PMG.

* [ZA] Change survey text on homepage

* [ZA] Add missing class to homepage featured MP image